rating inprogress

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+
 class CategoryController extends Controller
 {
     public function __construct()
@@ -13,28 +14,66 @@ class CategoryController extends Controller
         $this->middleware('auth');
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        $categories = Category::latest()->paginate(10);
+        $perPage = $request->perPage ?: 10;
+        $keyword = $request->keyword;
+
+        $categories = Category::query();
+
+        if ($keyword) {
+
+            $keyword = '%' . $keyword . '%';
+
+            $categories = $categories->where('title', 'like', $keyword);
+        }
+
+        $categories = $categories->latest()->paginate($perPage);
+
         return view('backend.pages.categories.index', compact('categories'));
     }
+
 
     public function create()
     {
         return view('backend.pages.categories.create');
     }
 
+
     public function store(Request $request)
     {
         $request->validate([
-            'category_name' => 'required|max:255|unique:categories,category_name',
-        ]);
-        Category::create([
-            'category_name' => $request->category_name
+            'title' => 'required',
+            'position' => 'required',
         ]);
 
-        return redirect()->back()->with('success', 'Category have been successfully created');
+        DB::beginTransaction();
+        try {
+
+
+            Category::create([
+                'title' => $request->title,
+                'slug' => $request->slug,
+                'position' => $request->position,
+            ]);
+
+            DB::commit();
+
+            return redirect()->back()->with('success', 'Category created successfully');
+
+        } catch (\Exception $exception) {
+            report($exception);
+            DB::rollBack();
+            return redirect()->back()->with('error', $exception->getMessage());
+        }
+
     }
+
+    /*    public function show(categories $categories)
+        {
+            return view('backend.categoriess.show', compact('categories'));
+        }*/
+
 
     public function edit(Category $category)
     {
@@ -45,18 +84,38 @@ class CategoryController extends Controller
     public function update(Request $request, Category $category)
     {
         $request->validate([
-            'category_name' => 'required|max:255|unique:categories,category_name,' . $category->id,
+            'title' => 'required',
+            'position' => 'required',
         ]);
-        $category->update([
-            'category_name' => $request->category_name
-        ]);
-        return redirect()->route('categories.index')->with('success', 'Category have been successfully updated');
+
+        DB::beginTransaction();
+        try {
+
+            $category->update([
+                'title' => $request->title,
+                'slug' => $request->slug,
+                'position' => $request->position,
+            ]);
+
+            DB::commit();
+
+            return redirect()->route('categories.index')->with('success', 'Category have been successfully updated');
+
+        } catch (\Exception $exception) {
+            report($exception);
+            DB::rollBack();
+            return redirect()->back()->with('error', $exception->getMessage());
+        }
+
     }
 
 
     public function destroy(Category $category)
     {
-        $category->delete();
-        return redirect()->back()->with('success', 'Category deleted success');
+        if ($category) {
+            $category->delete();
+            return redirect()->back()->with('success', 'Category have been successfully deleted');
+        }
+        abort(404);
     }
 }
